second commit

+++ src/main/java/com/programming/springblog/repository/PostRepository.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
-    List<Post> findByUser(User user);
+public interface PostRepository extends JpaRepository<Post, Long> {
+    
+	List<Post> findByUser(User user);
 }
